re-organises repo so that example code are part of testing framework
